+++ src/Compilers/CSharp/Test/Semantic/Semantics/UsingDeclarationTests.cs
@@ -287,23 +287,6 @@ class C
         }
 
         [Fact]
-        public void VariableCannotBeReAssignedAsUsingVariable()
-        {
-            var source = @"
-using System;
-class C
-{
-    static void Main()
-    {
-        var x = (IDisposable)null;
-        using var x2 = x;
-    }
-}
-";
-            CreateCompilation(source).VerifyDiagnostics();
-        }
-
-        [Fact]
         public void UsingVariableCannotBeUsedAsOutVariable()
         {
             var source = @"
@@ -350,6 +333,23 @@ class C
         }
 
         [Fact]
+        public void UsingVariableFromExistingVariable()
+        {
+            var source = @"
+using System;
+class C
+{
+    static void Main()
+    {
+        var x = (IDisposable)null;
+        using var x2 = x;
+    }
+}
+";
+            CreateCompilation(source).VerifyDiagnostics();
+        }
+
+        [Fact]
         public void UsingVariableFromExpression()
         {
             var source = @"
@@ -391,6 +391,58 @@ class C
 }
 ";
             CreateCompilation(source).VerifyDiagnostics();
+        }
+
+        [Fact]
+        public void UsingVariableInSwitchCase()
+        {
+            var source = @"
+using System;
+class C1 : IDisposable
+    {
+        public void Dispose() { }
+    }
+    class C2
+    {
+        public static void Main()
+        {
+            int x = 5;
+            switch (x)
+            {
+                case 5:
+                    using C1 o1 = new C1();
+                    break;
+            }
+        }
+    }";
+            CreateCompilation(source).VerifyDiagnostics(
+                // (15,21): error CS8389: A using variable cannot be used directly within a switch section (consider using braces). 
+                //                     using C1 o1 = new C1();
+                Diagnostic(ErrorCode.ERR_UsingVarInSwitchCase, "using C1 o1 = new C1();").WithLocation(15, 21)
+            );
+        }
+
+        [Fact]
+        public void UsingVariableDiagnosticsInDeclarationAreOnlyEmittedOnce()
+        {
+            var source = @"
+using System;
+class C1 : IDisposable
+{
+    public void Dispose() { }
+}
+class C2
+{
+    public static void Main()
+    {
+        using var c1 = new C1(), c2 = new C2();
+    }
+}";
+            CreateCompilation(source).VerifyDiagnostics(
+              // (11,15): error CS0819: Implicitly-typed variables cannot have multiple declarators
+              //         using (var c1 = new C1(), c2 = new C2())
+              Diagnostic(ErrorCode.ERR_ImplicitlyTypedVariableMultipleDeclarator, "var c1 = new C1(), c2 = new C2()").WithLocation(11, 15)
+            );
         }
     }
 }
